blob: split commitment dependencies into files
