+++ types/blob/powers_of_two.go
@@ -1,0 +1,28 @@
+package blob
+
+import (
+	"fmt"
+
+	"golang.org/x/exp/constraints"
+)
+
+// RoundDownPowerOfTwo returns the next power of two less than or equal to input.
+func RoundDownPowerOfTwo[I constraints.Integer](input I) (I, error) {
+	if input <= 0 {
+		return 0, fmt.Errorf("input %v must be positive", input)
+	}
+	roundedUp := RoundUpPowerOfTwo(input)
+	if roundedUp == input {
+		return roundedUp, nil
+	}
+	return roundedUp / 2, nil
+}
+
+// RoundUpPowerOfTwo returns the next power of two greater than or equal to input.
+func RoundUpPowerOfTwo[I constraints.Integer](input I) I {
+	var result I = 1
+	for result < input {
+		result = result << 1
+	}
+	return result
+}
