+++ types/blob/utils.go
@@ -1,0 +1,19 @@
+package blob
+
+import "bytes"
+
+// zeroPadIfNecessary pads the share with trailing zero bytes if the provided
+// share has fewer bytes than width. Returns the share unmodified if the
+// len(share) is greater than or equal to width.
+func zeroPadIfNecessary(share []byte, width int) (padded []byte, bytesOfPadding int) {
+	oldLen := len(share)
+	if oldLen >= width {
+		return share, 0
+	}
+
+	missingBytes := width - oldLen
+	padByte := []byte{0}
+	padding := bytes.Repeat(padByte, missingBytes)
+	share = append(share, padding...)
+	return share, missingBytes
+}
